// packages/ts.express.starter/src/features/config/types.ts
/*
    artifact generator: C:\My\wizzi\stfnbssl\wizzi.plugins\packages\wizzi.plugin.ts\lib\artifacts\ts\module\gen\main.js
    package: wizzi.plugin.ts@
    primary source IttfDocument: C:\My\wizzi\stfnbssl\wizzi.demo\packages\ts.express.starter\.wizzi\src\features\config\types.ts.ittf
    utc time: Thu, 29 Jun 2023 08:52:23 GMT
*/
export type ConfigType = { 
    port: number;
    sessionSecret: string;
    noCache: boolean;
    mongoHost: string;
    mongoUser: string;
    mongoPassword: string;
    mongoPath: string;
    mongoConnectUrl: string;
    isWizziDev: boolean;
    wizziBasePath: string;
    ittfPath: string;
    dataPath: string;
    metaHtmlIttfPath: string;
    metaFolderIttfPath: string;
    metaHtmlTextPath: string;
    jobsBasePath: string;
    userUserName: string;
    userDisplayName: string;
    userAvatarUrl: string;
};